requirements.txt
# install $ pip install -t . -r requrements.txt
python-dotenv
mysql-connector-python